осталась сортировка

+++ calculator/models.py
@@ -5,8 +5,8 @@ class Bank(models.Model):
     name = models.CharField(max_length=100, verbose_name='наименование банка')
     term_min = models.PositiveIntegerField(verbose_name='Срок ипотеки, ОТ')
     term_max = models.PositiveIntegerField(verbose_name='Срок ипотеки, ДО')
-    rate_min = models.PositiveSmallIntegerField(verbose_name='Ставка, ОТ')
-    rate_max = models.PositiveSmallIntegerField(verbose_name='Ставка, ДО')
+    rate_min = models.FloatField(verbose_name='Ставка, ОТ')
+    rate_max = models.FloatField(verbose_name='Ставка, ДО')
     payment_min = models.PositiveIntegerField(verbose_name='Сумма кредита, ОТ')
     payment_max = models.PositiveIntegerField(verbose_name='Сумма кредита, ДО')
